ai-answer-checker/requirements.txt
click
httpx
ruamel.yaml
pydantic
flask
sentence-transformers
scikit-learn
numpy